+++ CHANGELOG.md
@@ -9,4 +9,4 @@ This changelog references the relevant changes (new features, changes and bugs) 
   * Added localized title to datagrid.
   * Remove hard-coded paths for bpm forms.
   * Fixed account record widget position bug.
-  
+  * Added bpm localization variables to process start event.